Improve code - Refactor Get Item List - Refactor Service
