--- mixmax/Downloadable.swift
@@ -1,12 +1,0 @@
-//
-//  Downloadable.swift
-//  mixmax
-//
-//  Created by Apple on 4/23/18.
-//  Copyright © 2018 Vinh Nguyen. All rights reserved.
-//
-
-protocol Downloadable {
-    
-    var requestable : Requestable  { get set }
-}
