--- mixmax/DropboxItemDownload.swift
@@ -1,12 +1,0 @@
-//
-//  DropboxItemDownload.swift
-//  mixmax
-//
-//  Created by Apple on 4/23/18.
-//  Copyright © 2018 Vinh Nguyen. All rights reserved.
-//
-
-class DropboxItemDownload : ItemDownload {
-    
-    var downloader: Downloader = Downloader()
-}
