+++ mixmax/Parser.swift
@@ -9,8 +9,16 @@
 import Foundation
 
 protocol Parser {
-    var json: String { get }
-    func parse() -> [Item]
+    
+    var playableFiles: [String] { get }
+}
+
+extension Parser {
+    
+    var playableFiles: [String] {
+        
+        return ["mp3", "mp4", "wav", "wma", "m4v", "avi", "mpeg", "3gp", "m4a"]
+    }
 }
 
 struct Parser2<T: Item> {
