--- mixmax/Requestable.swift
@@ -1,12 +1,0 @@
-//
-//  Requestable.swift
-//  mixmax
-//
-//  Created by Apple on 4/23/18.
-//  Copyright © 2018 Vinh Nguyen. All rights reserved.
-//
-
-protocol Requestable {
-    
-    func request() -> Any
-}
